requirements.txt
Django~=2.2.10
django-allauth~=0.41.0
django-widget-tweaks~=1.4.8
Pillow~=7.2.0
stripe~=2.27.0